Moved more redundant single tests out of single tests directory. Also cleaned up some of the tests.
