+++ test/edu/umass/cs/gnsclient/client/singletests/simple/SelectClientTest.java
@@ -17,47 +17,45 @@
  *  Initial developer(s): Westy
  *
  */
-package edu.umass.cs.gnsclient.client.singletests.redundant;
+package edu.umass.cs.gnsclient.client.singletests.simple;
 
 import edu.umass.cs.gnsclient.client.GNSClientCommands;
+import edu.umass.cs.gnscommon.utils.RandomString;
 import edu.umass.cs.gnsclient.client.util.GuidEntry;
 import edu.umass.cs.gnsclient.client.util.GuidUtils;
-import edu.umass.cs.gnscommon.GNSProtocol;
-import edu.umass.cs.gnscommon.exceptions.client.ClientException;
-import edu.umass.cs.gnscommon.utils.RandomString;
 
+import edu.umass.cs.gnscommon.exceptions.client.ClientException;
 import edu.umass.cs.gnsserver.utils.DefaultGNSTest;
 import edu.umass.cs.utils.Utils;
 import java.io.IOException;
+import java.util.Arrays;
 
 import java.util.HashSet;
 import java.util.Set;
 import org.hamcrest.Matchers;
-
 import org.json.JSONArray;
 
 import org.json.JSONException;
 import org.junit.Assert;
-
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
 /**
- * Tests selectNear and selectWithin.
+ * Comprehensive functionality test for the GNS.
  *
  */
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SelectGeoTest extends DefaultGNSTest {
+public class SelectClientTest extends DefaultGNSTest {
 
-  private static GNSClientCommands clientCommands = null;
+  private static GNSClientCommands clientCommands;
   private static GuidEntry masterGuid;
   private static final Set<GuidEntry> createdGuids = new HashSet<>();
 
   /**
    *
    */
-  public SelectGeoTest() {
+  public SelectClientTest() {
     if (clientCommands == null) {
       try {
         clientCommands = new GNSClientCommands();
@@ -68,7 +66,7 @@ public class SelectGeoTest extends DefaultGNSTest {
       try {
         masterGuid = GuidUtils.getGUIDKeys(globalAccountName);
       } catch (Exception e) {
-        Utils.failWithStackTrace("Exception while looking up account guid: " + e);
+        Utils.failWithStackTrace("Exception when we were not expecting it: " + e);
       }
     }
   }
@@ -77,32 +75,29 @@ public class SelectGeoTest extends DefaultGNSTest {
    *
    */
   @Test
-  public void test_10_GeoSpatialSelectCreateGuids() {
+  public void test_01_testQuerySelect() {
+    String fieldName = "testQuery";
     try {
       for (int cnt = 0; cnt < 5; cnt++) {
-        GuidEntry testEntry = clientCommands.guidCreate(masterGuid,
-                "geoTest-" + RandomString.randomString(12));
+        GuidEntry testEntry = clientCommands.guidCreate(masterGuid, "queryTest-" + RandomString.randomString(12));
         createdGuids.add(testEntry); // save them so we can delete them later
-        clientCommands.setLocation(testEntry, 0.0, 0.0);
+        JSONArray array = new JSONArray(Arrays.asList(25));
+        clientCommands.fieldReplaceOrCreateList(testEntry, fieldName, array);
       }
     } catch (ClientException | IOException e) {
-      Utils.failWithStackTrace("Exception when we were not expecting it: " + e);
+      Utils.failWithStackTrace("Exception while trying to create the guids: " + e);
     }
-  }
+    waitSettle(100);
 
-  /**
-   *
-   */
-  @Test
-  public void test_20_GeoSpatialSelectSelectNear() {
     try {
-      JSONArray loc = new JSONArray();
-      loc.put(1.0);
-      loc.put(1.0);
-      JSONArray result = clientCommands.selectNear(GNSProtocol.LOCATION_FIELD_NAME.toString(), loc, 2000000.0);
+      String query = "~" + fieldName + " : ($gt: 0)";
+      JSONArray result = clientCommands.selectQuery(query);
+      for (int i = 0; i < result.length(); i++) {
+        System.out.println(result.get(i).toString());
+      }
       // best we can do should be at least 5, but possibly more objects in results
       Assert.assertThat(result.length(), Matchers.greaterThanOrEqualTo(5));
-    } catch (JSONException | ClientException | IOException e) {
+    } catch (ClientException | IOException | JSONException e) {
       Utils.failWithStackTrace("Exception executing selectNear: " + e);
     }
   }
@@ -111,30 +106,7 @@ public class SelectGeoTest extends DefaultGNSTest {
    *
    */
   @Test
-  public void test_30_GeoSpatialSelectSelectWithin() {
-    try {
-      JSONArray rect = new JSONArray();
-      JSONArray upperLeft = new JSONArray();
-      upperLeft.put(1.0);
-      upperLeft.put(1.0);
-      JSONArray lowerRight = new JSONArray();
-      lowerRight.put(-1.0);
-      lowerRight.put(-1.0);
-      rect.put(upperLeft);
-      rect.put(lowerRight);
-      JSONArray result = clientCommands.selectWithin(GNSProtocol.LOCATION_FIELD_NAME.toString(), rect);
-      // best we can do should be at least 5, but possibly more objects in results
-      Assert.assertThat(result.length(), Matchers.greaterThanOrEqualTo(5));
-    } catch (JSONException | ClientException | IOException e) {
-      Utils.failWithStackTrace("Exception executing selectWithin: " + e);
-    }
-  }
-
-  /**
-   *
-   */
-  @Test
-  public void test_40_GeoSpatialSelectCleanup() {
+  public void test_02_SelectCleanup() {
     try {
       for (GuidEntry guid : createdGuids) {
         clientCommands.guidRemove(masterGuid, guid.getGuid());
@@ -142,6 +114,16 @@ public class SelectGeoTest extends DefaultGNSTest {
       createdGuids.clear();
     } catch (ClientException | IOException e) {
       Utils.failWithStackTrace("Exception during cleanup: " + e);
+    }
+  }
+  
+  private static void waitSettle(long wait) {
+    try {
+      if (wait > 0) {
+        Thread.sleep(wait);
+      }
+    } catch (InterruptedException e) {
+      e.printStackTrace();
     }
   }
