+++ test/edu/umass/cs/gnsclient/client/singletests/simple/SingleReadArrayTest.java
@@ -17,13 +17,14 @@
  *  Initial developer(s): Westy
  *
  */
-package edu.umass.cs.gnsclient.client.singletests;
+package edu.umass.cs.gnsclient.client.singletests.simple;
 
 import edu.umass.cs.gnsclient.client.GNSClientCommands;
 import edu.umass.cs.gnsclient.client.util.GuidEntry;
 import edu.umass.cs.gnsclient.client.util.GuidUtils;
 import edu.umass.cs.gnscommon.exceptions.client.ClientException;
 import edu.umass.cs.gnscommon.utils.RandomString;
+import edu.umass.cs.gnscommon.utils.ThreadUtils;
 
 import edu.umass.cs.gnsserver.utils.DefaultGNSTest;
 import edu.umass.cs.utils.Utils;
@@ -36,11 +37,11 @@ import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
 /**
- * Test reads.
+ * Comprehensive functionality test for the GNS.
  *
  */
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class SingleReadTest extends DefaultGNSTest {
+public class SingleReadArrayTest extends DefaultGNSTest {
 
   private static GNSClientCommands clientCommands = null;
   private static GuidEntry masterGuid;
@@ -49,7 +50,7 @@ public class SingleReadTest extends DefaultGNSTest {
   /**
    *
    */
-  public SingleReadTest() {
+  public SingleReadArrayTest() {
     if (clientCommands == null) {
       try {
         clientCommands = new GNSClientCommands();
@@ -101,7 +102,7 @@ public class SingleReadTest extends DefaultGNSTest {
   @Test
   public void test_03_CreateField() {
     try {
-      clientCommands.fieldUpdate(subGuidEntry.getGuid(), "environment", "work", subGuidEntry);
+      clientCommands.fieldCreateOneElementList(subGuidEntry.getGuid(), "environment", "work", subGuidEntry);
     } catch (IOException | ClientException e) {
       Utils.failWithStackTrace("Exception during create field: " + e);
     }
@@ -111,10 +112,15 @@ public class SingleReadTest extends DefaultGNSTest {
    *
    */
   @Test
-  public void test_04_ReadField() {
+  public void test_04_ReadFieldTwice() {
     try {
       // read my own field
-      Assert.assertEquals("work", clientCommands.fieldRead(subGuidEntry.getGuid(), "environment", subGuidEntry));
+      Assert.assertEquals("work", clientCommands.fieldReadArrayFirstElement(subGuidEntry.getGuid(), "environment", subGuidEntry));
+      Assert.assertEquals("work", clientCommands.fieldReadArrayFirstElement(subGuidEntry.getGuid(), "environment", subGuidEntry));
+      ThreadUtils.sleep(5);
+      Assert.assertEquals("work", clientCommands.fieldReadArrayFirstElement(subGuidEntry.getGuid(), "environment", subGuidEntry));
+      ThreadUtils.sleep(5);
+      Assert.assertEquals("work", clientCommands.fieldReadArrayFirstElement(subGuidEntry.getGuid(), "environment", subGuidEntry));
     } catch (ClientException | IOException e) {
       Utils.failWithStackTrace("Exception reading field: " + e);
     }
